# backend/api/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: Database API for Nefryt LDS
  title: Nefryt LDS API
  version: 1.0.0
servers:
- description: Test server
  url: http://192.168.1.231:8080/
- description: Developer's local server
  url: http://localhost:8080/
security:
- bearerAuth: []
paths:
...
  /trend:
    get:
      description: List all trends
      operationId: list_trends
      parameters:
        - description: Page number
          in: query
          name: $page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - description: Page size
          in: query
          name: $size
          required: false
          schema:
            type: integer
            maximum: 10000
            minimum: 1
            default: 50
        - description: Query filter in OData standard
          explode: true
          in: query
          name: $filter
          required: false
          schema:
            type: string
          style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Trend'
                type: array
          description: A array of trends
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List trends
      tags:
      - trends
      x-openapi-router-controller: api.controllers.trends_controller
    post:
      description: Create a trend
      operationId: create_trend
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Trend'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trend'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create trend
      tags:
      - trends
      x-openapi-router-controller: api.controllers.trends_controller
...
  /trend/{trendId}:
    delete:
      description: Deletes specific trend
      operationId: delete_trend_by_id
      parameters:
      - description: The id of the trend to retrieve
        explode: false
        in: path
        name: trendId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: No content
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Deletes trend
      tags:
      - trends
      x-openapi-router-controller: api.controllers.trends_controller
    get:
      description: Info for specific trend
      operationId: get_trend_by_id
      parameters:
      - description: The id of the trend to retrieve
        explode: false
        in: path
        name: trendId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trend'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Detail trend
      tags:
      - trends
      x-openapi-router-controller: api.controllers.trends_controller
    put:
      description: Update a trend
      operationId: update_trend
      parameters:
      - description: The id of the trend to retrieve
        explode: false
        in: path
        name: trendId
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Trend'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trend'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Update trend
      tags:
      - trends
      x-openapi-router-controller: api.controllers.trends_controller
...
  /trend_def:
    get:
      description: List all trend defs
      operationId: list_trend_defs
      parameters:
        - description: Page number
          in: query
          name: $page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - description: Page size
          in: query
          name: $size
          required: false
          schema:
            type: integer
            maximum: 10000
            minimum: 1
            default: 50
        - description: Query filter in OData standard
          explode: true
          in: query
          name: $filter
          required: false
          schema:
            type: string
          style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TrendDef'
                type: array
          description: A array of trend defs
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List trend defs
      tags:
      - trends
      x-openapi-router-controller: api.controllers.trends_controller
  /template:
    get:
      description: List all templates
      operationId: list_templates
      parameters:
        - description: Page number
          in: query
          name: $page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - description: Page size
          in: query
          name: $size
          required: false
          schema:
            type: integer
            maximum: 10000
            minimum: 1
            default: 50
        - description: Query filter in OData standard
          explode: true
          in: query
          name: filter
          required: false
          schema:
            type: string
          style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Template'
                type: array
          description: An array of templates
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: List templates
      tags:
        - templates
    post:
      description: Create a template
      operationId: create_template
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Template'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Create template
      tags:
        - templates
  /template/{templateId}:
    delete:
      description: Deletes specific template
      operationId: delete_template_by_id
      parameters:
        - description: The id of the template to retrieve
          explode: false
          in: path
          name: templateId
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "204":
          description: No content
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Delete template
      tags:
        - templates
    get:
      description: Info for specific template
      operationId: get_template_by_id
      parameters:
        - description: The id of the template to retrieve
          explode: false
          in: path
          name: templateId
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Get template details
      tags:
        - templates
    put:
      description: Updates a template
      operationId: update_template
      parameters:
        - description: The id of the template to retrieve
          explode: false
          in: path
          name: templateId
          required: true
          schema:
            type: integer
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Template'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Updates templates
      tags:
        - templates
...
components:
...
  schemas:
    Axis:
      properties:
        TrendsID:
          items:
            type: integer
          type: array
          default: [ ]
        Title:
          type: string
        Unit:
          type: string
        ScaledMin:
          type: number
        ScaledMax:
          type: number
      type: object
      required:
        - Unit
        - ScaledMin
        - ScaledMax
...
    Template:
      properties:
        ID:
          type: integer
          nullable: false
        Name:
          type: string
        Axes:
          items:
            $ref: '#/components/schemas/Axis'
          type: array
          default: [ ]
      required:
        - Name
        - ID
      type: object
    Trend:
      example:
        TrendGroupID: 0
        Format: Format
        Symbol: Symbol
        Color: Color
        NodeID: 6
        TrendDefID: TrendDefID
        ID: 0
        TimeExponent: 1
        Unit: Unit
        Name: Name
        RawMin: 0
        RawMax: 10000
        ScaledMin: -0.5
        ScaledMax: 1.6
      properties:
        ID:
          description: none
          example: 0
          nullable: false
          readOnly: true
          title: ID
          type: integer
        Name:
          description: none
          pattern: ^.{0,30}
          title: Name
          type: string
        TrendGroupID:
          description: none
          title: TrendGroupID
          type: integer
        TrendDefID:
          description: none
          title: TrendDefID
          type: string
        NodeID:
          description: none
          title: NodeID
          type: integer
        TimeExponent:
          description: none
          title: TimeExponent
          type: integer
        Format:
          description: none
          title: Format
          type: string
        Unit:
          description: none
          title: Unit
          type: string
        Color:
          title: Color
          type: string
        Symbol:
          title: Symbol
          type: string
        RawMin:
          description: none
          nullable: false
          title: RawMin
          type: integer
        RawMax:
          description: none
          nullable: false
          title: RawMax
          type: integer
        ScaledMin:
          description: none
          nullable: false
          title: ScaledMin
          type: number
        ScaledMax:
          description: none
          nullable: false
          title: RawMin
          type: number
      required:
      - TrendDefID
      - RawMin
      - RawMax
      - ScaledMin
      - ScaledMax
      title: Trend
      type: object
...
    TrendDef:
      example:
        ID: "0"
        Name: Name
      properties:
        ID:
          description: none
          example: "0"
          nullable: false
          readOnly: true
          title: ID
          type: string
        Name:
          description: none
          pattern: ^.{0,30}
          title: Name
          type: string
      required:
      - ID
      - Name
      title: TrendDef
      type: object
...
    Error:
      properties:
        code:
          format: int32
          title: code
          type: integer
        message:
          title: message
          type: string
      required:
      - code
      - message
      title: Error
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http